Creating header file for physics ecs
